better suspense model
